pyproject.toml
[tool.poetry]
name = "EnteRAG"
version = "0.1.0"
description = "AI-powered Enterprise RAG"
authors = ["Fei(Michael) Wang <numerical.ocean@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.103.0"
uvicorn = "0.23.2"
pydantic = "^2.0.0"
requests = "^2.22.0"
loguru = "^0.7.0"
joblib = "^1.2.0"
scikit-learn = "^1.2.0"
jinja2 = "^3.1.4"
optimum = {extras = ["exporters", "onnxruntime"], version = "^1.19.2"}
python-dotenv = "^1.0.1"
sentence-transformers = "^2.7.0"
elasticsearch = "^8.13.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.2"
black = {version = "^24.3", allow-prereleases = true}
autopep8 = "^2.0.0"
ipdb = "^0.13.0"
pylint = "^3.0.0"
pyspark = "^3.5.1"

[tool.poetry.group.aws]
optional = true

[tool.poetry.group.aws.dependencies]
mangum = "^0.17.0"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
exclude = '''
/(
  \.toml
  |\.sh
  |\.git
  |\.ini
  |Dockerfile
  |Jenkinfile
)/
'''
